+++ app/cdap/components/GridTable/components/Typography/types.ts
